requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.12.0
    # via
    #   tensorflow-datasets
    #   tensorflow-metadata
argon2-cffi==20.1.0
    # via notebook
async-generator==1.10
    # via nbclient
attrs==20.3.0
    # via
    #   jsonschema
    #   tensorflow-datasets
augmentor==0.2.8
    # via -r requirements.in
backcall==0.2.0
    # via ipython
bleach==3.3.0
    # via nbconvert
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
cffi==1.14.5
    # via argon2-cffi
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   pip-tools
    #   wandb
configparser==5.0.2
    # via wandb
cycler==0.10.0
    # via matplotlib
decorator==4.4.2
    # via
    #   ipython
    #   moviepy
    #   networkx
defusedxml==0.7.1
    # via nbconvert
dill==0.3.3
    # via tensorflow-datasets
docker-pycreds==0.4.0
    # via wandb
entrypoints==0.3
    # via nbconvert
enum34==1.1.10
    # via pandas-ml
future==0.18.2
    # via
    #   augmentor
    #   tensorflow-datasets
gitdb==4.0.7
    # via gitpython
gitpython==3.1.14
    # via wandb
googleapis-common-protos==1.53.0
    # via tensorflow-metadata
gviz-api==1.9.0
    # via tensorboard-plugin-profile
idna==2.10
    # via requests
imageio-ffmpeg==0.4.3
    # via moviepy
imageio==2.9.0
    # via
    #   imgaug
    #   moviepy
    #   scikit-image
imgaug==0.4.0
    # via -r requirements.in
importlib-resources==5.1.2
    # via tensorflow-datasets
ipykernel==5.5.3
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   notebook
    #   qtconsole
    #   traitlets
ipython==7.22.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==7.6.3
    # via jupyter
jedi==0.18.0
    # via ipython
jinja2==2.11.3
    # via
    #   nbconvert
    #   notebook
joblib==1.0.1
    # via scikit-learn
jsonschema==3.2.0
    # via nbformat
jupyter-client==6.1.12
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.0
    # via jupyter
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter==1.0.0
    # via -r requirements.in
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.0
    # via ipywidgets
kiwisolver==1.3.1
    # via matplotlib
markupsafe==1.1.1
    # via jinja2
matplotlib==3.4.1
    # via
    #   -r requirements.in
    #   imgaug
    #   scikit-image
    #   seaborn
mistune==0.8.4
    # via nbconvert
moviepy==1.0.3
    # via -r requirements.in
mypy-extensions==0.4.3
    # via typing-inspect
nbclient==0.5.3
    # via nbconvert
nbconvert==6.0.7
    # via
    #   jupyter
    #   notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via nbclient
networkx==2.5.1
    # via scikit-image
notebook==6.3.0
    # via
    #   jupyter
    #   widgetsnbextension
numpy==1.19.5
    # via
    #   -r requirements.in
    #   augmentor
    #   imageio
    #   imgaug
    #   matplotlib
    #   moviepy
    #   opencv-python
    #   pandas
    #   pywavelets
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tensorflow-datasets
    #   tifffile
opencv-python==4.5.1.48
    # via
    #   -r requirements.in
    #   imgaug
packaging==20.9
    # via bleach
git+https://github.com/EgorDm/pandas-ml.git
    # via -r requirements.in
pandas==1.2.4
    # via
    #   -r requirements.in
    #   pandas-ml
    #   seaborn
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.2
    # via jedi
pathtools==0.1.2
    # via wandb
pep517==0.10.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.2.0
    # via
    #   augmentor
    #   imageio
    #   imgaug
    #   matplotlib
    #   scikit-image
git+https://github.com/jazzband/pip-tools.git
    # via -r requirements.in
proglog==0.1.9
    # via moviepy
prometheus-client==0.10.1
    # via notebook
promise==2.3
    # via
    #   tensorflow-datasets
    #   wandb
prompt-toolkit==3.0.18
    # via
    #   ipython
    #   jupyter-console
protobuf==3.15.8
    # via
    #   googleapis-common-protos
    #   tensorboard-plugin-profile
    #   tensorflow-datasets
    #   tensorflow-metadata
    #   wandb
psutil==5.8.0
    # via wandb
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pycparser==2.20
    # via cffi
pydot==1.4.2
    # via -r requirements.in
pygments==2.8.1
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
    #   pydot
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
    #   wandb
pytz==2021.1
    # via pandas
pywavelets==1.1.1
    # via scikit-image
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   wandb
pyzmq==22.0.3
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.0.3
    # via jupyter
qtpy==1.9.0
    # via qtconsole
requests==2.25.1
    # via
    #   moviepy
    #   tensorflow-datasets
    #   wandb
scikit-image==0.18.1
    # via imgaug
scikit-learn==0.24.1
    # via
    #   pandas-ml
    #   sklearn
scipy==1.6.2
    # via
    #   -r requirements.in
    #   imgaug
    #   scikit-image
    #   scikit-learn
    #   seaborn
seaborn==0.11.1
    # via -r requirements.in
send2trash==1.5.0
    # via notebook
sentry-sdk==1.0.0
    # via wandb
shapely==1.7.1
    # via imgaug
shortuuid==1.0.1
    # via wandb
simple-parsing==0.0.14.post1
    # via -r requirements.in
six==1.15.0
    # via
    #   absl-py
    #   argon2-cffi
    #   bleach
    #   cycler
    #   docker-pycreds
    #   gviz-api
    #   imgaug
    #   jsonschema
    #   promise
    #   protobuf
    #   python-dateutil
    #   tensorboard-plugin-profile
    #   tensorflow-datasets
    #   wandb
sklearn==0.0
    # via -r requirements.in
smmap==4.0.0
    # via gitdb
subprocess32==3.5.4
    # via wandb
tensorboard-plugin-profile==2.4.0
    # via -r requirements.in
tensorflow-datasets==4.2.0
    # via -r requirements.in
tensorflow-metadata==0.29.0
    # via tensorflow-datasets
termcolor==1.1.0
    # via tensorflow-datasets
terminado==0.9.4
    # via notebook
testpath==0.4.4
    # via nbconvert
threadpoolctl==2.1.0
    # via scikit-learn
tifffile==2021.4.8
    # via scikit-image
toml==0.10.2
    # via pep517
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.60.0
    # via
    #   -r requirements.in
    #   augmentor
    #   moviepy
    #   proglog
    #   tensorflow-datasets
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==3.7.4.3
    # via typing-inspect
typing-inspect==0.6.0
    # via simple-parsing
urllib3==1.26.4
    # via
    #   requests
    #   sentry-sdk
wandb==0.10.26
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
werkzeug==1.0.1
    # via tensorboard-plugin-profile
widgetsnbextension==3.5.1
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
